ADAPTED VERSION OF L-PY
This version does not include the turtle interpretation and visualization aspects, thus dependencies on PlantGL lib have been removed.

Note that older changes have not been tracket by git.
This version is not intended to be upstream compatible with the official L-Py version.

For further information refer to README.txt